# Conflicts:
#	elichika/elichika/chainer2onnx.py
#	elichika/elichika/parser/core.py
#	elichika/elichika/parser/functions_builtin.py

+++ compiler/passes.cc
@@ -7,15 +7,20 @@
 #include <compiler/config.h>
 #include <compiler/constant_propagation.h>
 #include <compiler/flags.h>
+#include <compiler/flops.h>
 #include <compiler/fusion.h>
 #include <compiler/gradient.h>
 #include <compiler/graph.h>
+#include <compiler/memory_simulator.h>
 #include <compiler/model.h>
-#include <compiler/recompute.h>
 #include <compiler/scheduler.h>
+#include <compiler/shape_evaluator.h>
 #include <compiler/simplifier.h>
 #include <compiler/subgraph_canonicalizer.h>
 #include <compiler/type_inference.h>
+
+#include <compiler/computation_order/core.h>
+#include <compiler/gradient_with_order.h>
 
 namespace chainer_compiler {
 
@@ -51,6 +56,9 @@ void RunDefaultPasses(Model* model, bool gen_backprop) {
 }
 
 void RunDefaultPasses(Graph* graph, bool gen_backprop) {
+    // TODO(hamaji): Improve backend selection probably by `CompilerConfig`.
+    g_modify_pool_with_imbalanced_pads = !g_use_ngraph;
+
     std::unique_ptr<CompilerConfig> ccfg{GetCompilerConfig(g_backend_name)};
 
     InferAllDtypeAndShape(graph);
@@ -72,20 +80,36 @@ void RunDefaultPasses(Graph* graph, bool gen_backprop) {
 
     Recursively(PropagateConstants, graph);
 
+    Recursively(EvaluateShapes, graph);
+
     Recursively([](Graph* g) { g->DeleteDetached(); }, graph);
 
     dump_onnx(g_dump_after_simplification, "after simplification");
 
-    if (gen_backprop) AddGradientNodesForTraining(graph);
+    bool skip_scheduling = false;
+    if (gen_backprop) {
+        if (g_computation_order.empty()) {
+            // normal computation order
+            AddGradientNodesForTraining(graph);
+        } else {
+            // specified computation order
+            skip_scheduling = true;
+            auto orders = GetComputationOrder(*graph, g_computation_order);
+            AddGradientNodesForTrainingWithOrders(graph, orders);
+            // SimplifyOps({Node::kIdentity}, graph);
+        }
+    }
 
     // TODO(hamaji): Make it possible to infer shapes here.
     // if (!g_skip_inference) graph->InferShapes();
 
-    Recursively([&ccfg, gen_backprop](Graph* g) { Simplify(*ccfg, g, gen_backprop); }, graph);
+    if (!skip_scheduling) {
+        Recursively([&ccfg, gen_backprop](Graph* g) { Simplify(*ccfg, g, gen_backprop); }, graph);
 
-    Recursively(PropagateConstants, graph);
+        Recursively(PropagateConstants, graph);
 
-    Recursively([](Graph* g) { g->DeleteDetached(); }, graph);
+        Recursively([](Graph* g) { g->DeleteDetached(); }, graph);
+    }
 
     dump_onnx(g_dump_after_gradient, "after gradient generation");
 
@@ -93,19 +117,24 @@ void RunDefaultPasses(Graph* graph, bool gen_backprop) {
         graph->DumpSubGraphs();
     }
 
-    if (g_recompute_relu) GetReluRecompute(graph, g_recompute_relu);
-
-    if (g_fuse_operations) {
-        FuseOperations(graph, g_use_tvm);
-        dump_onnx(g_dump_after_fusion, "after fusion");
+    if (!skip_scheduling) {
+        if (g_fuse_operations) {
+            FuseOperations(graph, g_use_tvm, g_use_ngraph);
+            dump_onnx(g_dump_after_fusion, "after fusion");
+        }
     }
 
     int64_t order = 0;
     Recursively([&order](Graph* g) { order = ScheduleComputation(*g, order); }, graph);
 
-    dump_onnx(g_dump_after_scheduling, "after scheduling");
+    if (g_compiler_log) {
+        ShowSimulatedMemoryUsage(*graph);
+        ShowFlops(*graph);
+    }
 
     Recursively(CollectGarbageNode, graph);
+
+    dump_onnx(g_dump_after_scheduling, "after scheduling");
 
     Recursively([&ccfg](Graph* g) { CheckAllOpsSupported(*ccfg, g); }, graph);
 }
